+++ src/NBT/NBTGZipFile.cpp
@@ -9,7 +9,6 @@ namespace voxel
 	NBTGZipFile::NBTGZipFile(std::string name)
 	: name(name)
 	{
-		//Empty
 	}
 
 	void NBTGZipFile::load()
@@ -26,25 +25,25 @@ namespace voxel
 	{
 		if (!this->ostream.open(this->name))
 			throw NBTException("Failed to open file " + this->name);
-		for (uint32_t i = 0; i < this->tags.size(); ++i)
+		for (size_t i = 0; i < this->tags.size(); ++i)
 			this->tags[i]->writeData(this);
 		this->ostream.close();
 	}
 
 	void NBTGZipFile::printDebug()
 	{
-		for (uint32_t i = 0; i < this->tags.size(); ++i)
+		for (size_t i = 0; i < this->tags.size(); ++i)
 			this->tags[i]->printDebug();
 	}
 
 	bool NBTGZipFile::read(void *data, size_t len)
 	{
-		return (this->istream.read(data, len) == (ssize_t)len);
+		return this->istream.read(data, len) == (ssize_t)len;
 	}
 
 	bool NBTGZipFile::write(void *data, size_t len)
 	{
-		return (this->ostream.write(data, len) == (ssize_t)len);
+		return this->ostream.write(data, len) == (ssize_t)len;
 	}
 
 }
